Module ic_nns_common::registry

Constants

SUBNET_LIST_KEY

Functions

encode_or_panic
Wraps around Message::encode and panics on error.

get_icp_xdr_conversion_rate_record

get_latest_version
Returns the latest version of the registry

get_subnet_list_record
Gets the subnet list record.

get_subnet_record

get_value
Returns the deserialized value associated with the given key and version. If the version is None, then the "latest" version is returned.

mutate_registry
Tries to mutate the registry. If it succeeds, returns the version at which that mutation happened.