Function ic_nns_common::registry::get_value

pub async fn get_value<'_, T: Message + Default>(
    key: &'_ [u8], 
    version: Option<u64>
) -> Result<(T, u64), Error>

Returns the deserialized value associated with the given key and version. If the version is None, then the "latest" version is returned.

The returned tuple is (value, version) with "version" being the version at which the value corresponding to the key was last mutated (inserted, updated, or deleted). This function will propagate any error from the deserialization or decoding.